+++ .gitignore
@@ -63,3 +63,6 @@ ecomap/www/uploads/
 
 # Vagrant files
 .vagrant/
+
+# idea
+.idea/
