+++ page_object_ecomap/tests/test_data_file.txt
@@ -1,0 +1,3 @@
+base_url: http://localhost
+email: admin@ecomap.com
+password: secre!
